requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
csvw==3.5.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
decorator==5.2.1
dlinfo==2.0.0
docopt==0.6.2
dotenv==0.9.9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
espeakng-loader==0.2.4
executing==2.2.0
filelock==3.18.0
flask==3.1.1
frozenlist==1.7.0
fsspec==2025.7.0
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
ipython==9.4.0
ipython-pygments-lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jaraco-context==6.0.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kokoro==0.9.4
langcodes==3.5.0
language-data==1.3.0
language-tags==1.2.0
loguru==0.7.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
misaki==0.9.4
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
murmurhash==1.0.13
networkx==3.5
num2words==0.5.14
numpy==2.3.2
oauthlib==3.3.1
ollama==0.5.1
packaging==25.0
parso==0.8.4
pexpect==4.9.0
phonemizer-fork==3.3.2
pip==24.0
preshed==3.0.10
prompt-toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyaudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyttsx3==2.99
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.1
rdflib==7.1.4
referencing==0.36.2
regex==2025.7.31
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339==6.2
rfc3986==1.5.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scipy==1.16.1
segments==2.3.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.3.0.post1
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechrecognition==3.14.3
srsly==2.5.1
stack-data==0.6.3
sympy==1.14.0
thinc==8.3.6
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
werkzeug==3.1.3
wikipedia==1.4.0
wolframalpha==5.1.3
wrapt==1.17.2
xmltodict==0.14.2
yarl==1.20.1
